Refactor 'resolver' classes into separate package
